+++ addon/utils/snapshot-transform.js
@@ -1,5 +1,5 @@
 export default {
-  transformForSerialize: function(snapshot, skipUnchangedAttrs = true) {
+  transformForSerialize: function (snapshot, skipUnchangedAttrs = true) {
     if (skipUnchangedAttrs) {
       let changedAttributes = Object.keys(snapshot.changedAttributes());
       for (let attrKey in snapshot._attributes) {
@@ -9,12 +9,16 @@ export default {
         }
       }
 
-      snapshot.eachAttribute = function(callback, binding) {
-        snapshot.record.eachAttribute(function(name, meta) {
-          if (name in snapshot._attributes) {
-            callback.call(binding, name, meta);
-          }
-        }, binding);
+      snapshot.eachAttribute = function (callback, binding) {
+        if (!snapshot.record) {
+          return snapshot;
+        } else {
+          snapshot.record.eachAttribute(function (name, meta) {
+            if (name in snapshot._attributes) {
+              callback.call(binding, name, meta);
+            }
+          }, binding);
+        }
       };
     }